# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Бухгалтер API
  description: |
    API приложения «Бухгалтер» для учёта личных финансов.

    Ответы с ошибками (`4xx`, `5xx`) используют объект `error` с полями `code`, `message` и опционально
    **`field`** — имя поля тела запроса для подсветки в форме UI (например `monthly_payment`, `term_months`).
    Схема: `ErrorResponse` → `error.field`.

    **Серверный кеш (in-memory):** успешные ответы `GET` кешируются в процессе сервера (справочники — 5 мин,
    остальное — 1 мин); при любой мутации кеш пользователя сбрасывается.
    **Клиент (web / Android):** SWR ref-cache в `localStorage` для `GET /api/v1/*` (кроме health, export, preview, version, `GET /setup/status`).
    Подробнее — [ui-api-cache.md](../ui-api-cache.md).

    **Счета (UI):** на `/accounts` меню «⋯» вызывает inline-редактирование (`PUT /accounts/{id}`),
    `PUT /accounts/{id}/credit-limit` (кредитные карты), `POST /accounts/{id}/primary`,
    `POST /accounts/{id}/archive`, `DELETE /accounts/{id}`.
    На `/accounts/{id}` — та же форма редактирования в шапке. Поля: название, банк, начальный баланс.
    Лимит кредитной карты меняется только через `PUT /accounts/{id}/credit-limit` (см. [ui-credit-cards.md](../ui-credit-cards.md)).
    В форме «Начальный баланс» — значение `initial_balance`, не текущий `balance_display` (см. [data-model.md](../data-model.md)).
    Подробнее — [ui-row-actions.md](../ui-row-actions.md).
  version: 1.5.0
servers:
  - url: /api/v1
paths:
...
  /setup:
    post:
      summary: Начальная настройка (один раз)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - admin_login
                - admin_display_name
                - admin_password
                - admin_password_confirm
              properties:
                admin_login:
                  type: string
                  minLength: 3
                  maxLength: 32
                  description: Логин администратора
                admin_display_name:
                  type: string
                  minLength: 1
                  maxLength: 64
                  description: Имя администратора (display_name)
                admin_password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с admin_login
                admin_password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с admin_password
                registration_enabled:
                  type: boolean
                  default: false
                external_url:
                  type: string
                  description: Опционально; URL за reverse proxy (http/https)
      responses:
        '201':
          description: Настройка завершена
        '400':
          description: Ошибка валидации (пароли не совпадают, пустое имя и т.д.)
        '409':
          description: Настройка уже выполнена
...
  /auth/register:
    post:
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
                - password_confirm
              properties:
                login:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с login
                password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с password
                display_name:
                  type: string
                  description: Опционально; по умолчанию — login
      responses:
        '201':
          description: Пользователь создан (статус pending, сессия не выдаётся)
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации (пароли не совпадают и т.д.)
        '403':
          description: Регистрация отключена
...
  /admin/settings:
    get:
      summary: Системные настройки
      description: |
        Внешний URL и признак заданного ключа уведомлений.
        Самостоятельная регистрация и модули — `GET/PUT /admin/features` (флаг `registration`).
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Настройки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSettings'
    put:
      summary: Обновить системные настройки
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSettingsUpdate'
      responses:
        '200':
          description: Настройки обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSettings'
        '400':
          description: Ошибка валидации external_url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                external_url:
                  $ref: '#/components/examples/ApiErrorExternalURL'
  /admin/features:
    get:
      summary: Каталог флагов модулей
      description: Список известных флагов с метаданными и текущим состоянием (только admin).
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Каталог
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminFeatureFlag'
    put:
      summary: Обновить флаги модулей
      description: |
        Частичная или полная запись overrides. Неизвестный ключ — 400.
        При выключенном модуле API модуля отвечает 404 с кодом `FEATURE_DISABLED`.
        Значение `registration` также отдаётся публично в `GET /setup/status` как `registration_enabled`.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminFeatureFlagsUpdate'
      responses:
        '200':
          description: Актуальный снимок флагов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagsSnapshot'
        '400':
          description: Неизвестный ключ или пустое тело
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                unknown_key:
                  $ref: '#/components/examples/ApiErrorUnknownFeatureFlag'
                empty:
                  $ref: '#/components/examples/ApiErrorFeaturesEmpty'
...
  /admin/users:
    get:
      summary: Список пользователей
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Пользователи
    post:
      summary: Создать пользователя
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
                - password_confirm
              properties:
                login:
                  type: string
                  minLength: 3
                  maxLength: 32
                password:
                  type: string
                  minLength: 8
                  description: Минимум 8 символов, хотя бы одна буква и одна цифра, не совпадает с login
                password_confirm:
                  type: string
                  minLength: 8
                  description: Должен совпадать с password
                display_name:
                  type: string
                  description: Опционально; по умолчанию — login
                is_admin:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Пользователь создан
        '400':
          description: Ошибка валидации (пароли не совпадают и т.д.)
...
  /version/check:
    get:
      summary: Проверка доступного обновления
      description: |
        Сравнивает текущую версию приложения с последним релизом на GitHub.
        Ответ кешируется на сервере 24 часа. Веб-клиент запрашивает endpoint не чаще одного раза в сутки.
        Модальное предупреждение (бэкап, release notes, ссылка на **Настройки → Админка → Бекапы**) показывается
        **только администраторам** при `update_available: true`.
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionCheckResponse'
              example:
                current_version: 1.2.1
                latest_version: 1.2.2
                update_available: true
                release_url: https://github.com/kai-zer-ru/buhgalter/releases/tag/v1.2.2
...
  /categories:
    get:
      summary: Список категорий
      description: |
        Сервер кеширует ответ in-memory (TTL 5 мин); кеш сбрасывается при POST/PUT/DELETE
        категорий и подкатегорий, reorder и set primary.
      security:
        - sessionCookie: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
      responses:
        '200':
          description: Список категорий
    post:
      summary: Создать категорию
      security:
        - sessionCookie: []
      responses:
        '201':
          description: Категория создана
...
  /merchants:
    get:
      summary: Список магазинов
      description: |
        Справочник магазинов пользователя. Кеш GET TTL 5 мин; сброс при мутациях.
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Merchant'
    post:
      summary: Создать магазин
      description: Get-or-create по имени (без учёта регистра, включая кириллицу).
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantUpsert'
      responses:
        '201':
          description: Создан или найден существующий
...
  /tags:
    get:
      summary: Список тегов
      description: |
        Опциональный фильтр `q` — подстрока в имени. Кеш GET TTL 5 мин.
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Создать тег
      description: Get-or-create по имени.
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NameCreate'
      responses:
        '201':
          description: Создан или найден
...
  /transaction-templates:
    get:
      summary: Список шаблонов операций
      description: |
        Именованные пресеты для быстрого создания операций. Кеш GET TTL 5 мин.
        Сортировка: `sort_order`, затем имя.
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionTemplate'
    post:
      summary: Создать шаблон операции
      description: |
        Не создаёт операцию. Лимит — 100 шаблонов на пользователя.
        Для `transfer` обязательны разные `account_id` и `to_account_id`.
        `amount` можно не передавать (сумма при создании операции).
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTemplateUpsert'
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTemplate'
        '400':
          description: Ошибка валидации или лимит
...
  /debtors:
    get:
      summary: Список должников
      description: |
        Сервер кеширует ответ in-memory (TTL 5 мин); кеш сбрасывается при `POST /debtors`.
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Список
    post:
      summary: Создать должника
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '201':
          description: Создан
...
  /import/jobs:
    post:
      summary: Импорт операций в фоне (async job)
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
          description: Повтор с тем же ключом возвращает ту же задачу/результат без дублей
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                preset:
                  type: string
                  enum: [cubux, custom]
                deduplicate:
                  type: boolean
                column_map:
                  type: string
                account_map:
                  type: string
                category_map:
                  type: string
                auto_subcategory:
                  type: boolean
                subcategory_map:
                  type: string
      responses:
        '202':
          description: Задача импорта создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportJob'
...
  /export:
    get:
      summary: Экспорт операций в Cubux CSV
      security:
        - sessionCookie: []
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: account_id
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: CSV с BOM
          content:
            text/csv:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
...
  examples:
...
    ApiErrorUnknownFeatureFlag:
      summary: Неизвестный ключ флага модуля
      value:
        error:
          code: VALIDATION_ERROR
          message: unknown feature flag
          field: features
    ApiErrorFeaturesEmpty:
      summary: Пустое тело обновления флагов
      value:
        error:
          code: VALIDATION_ERROR
          message: Укажите хотя бы один флаг
...
    ApiErrorExternalURL:
      summary: Некорректный external_url
      value:
        error:
          code: VALIDATION_ERROR
          message: external_url должен начинаться с http:// или https://
...
  schemas:
...
    VersionCheckResponse:
      type: object
      required: [current_version, update_available]
      properties:
        current_version:
          type: string
          description: Версия текущей установки
        latest_version:
          type: string
          description: Последняя версия на GitHub (если удалось получить)
        update_available:
          type: boolean
          description: true, если latest_version новее current_version
        release_url:
          type: string
          format: uri
          description: Ссылка на страницу релиза на GitHub
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Код ошибки: UNAUTHORIZED, FORBIDDEN, INVALID_CREDENTIALS, REGISTRATION_DISABLED,
                FEATURE_DISABLED, USER_PENDING_MODERATION, USER_BANNED, USER_STATUS_INVALID,
                USER_STATUS_TRANSITION, RATE_LIMITED, VALIDATION_ERROR, INTERNAL_ERROR, NOT_FOUND,
                CONFLICT, ALREADY_CONFIGURED, SERVICE_UNAVAILABLE.
                Для смены пароля и большинства 400 — `VALIDATION_ERROR`; текст в `message`
                локализуется по ключам (PASSWORDS_MISMATCH, ERR_CHANNEL, ERR_SYSTEM_CATEGORY_PLANNED,
                ERR_TX_TYPE_CHANGE, ERR_CREDIT_INVALID_PAYMENT, ERR_CREDIT_INVALID_TERM, ERR_CREDIT_INVALID_MORTGAGE, …).
            message:
              type: string
              description: Локализованное сообщение (заголовок Accept-Language)
            field:
              type: string
              description: |
                Опциональное имя поля запроса для привязки ошибки в форме (например `monthly_payment`,
                `term_months`, `debit_time_local`, `credit_kind`, `property_price`, `body`).
                Заполняется для ключей `ERR_CREDIT_*`, `ERR_INVALID_*`, `ERR_ACCOUNT_*` и аналогичных.
      example:
        error:
          code: VALIDATION_ERROR
          message: Некорректные параметры запроса
          field: monthly_payment
    User:
      type: object
      required: [id, login, display_name, is_admin, status, language, currency, timezone, theme]
      properties:
        id:
          type: string
        login:
          type: string
        display_name:
          type: string
        is_admin:
          type: boolean
        status:
          type: string
          enum: [active, pending, banned]
        language:
          type: string
        currency:
          type: string
        timezone:
          type: string
        theme:
          type: string
          enum: [light, dark, system]
          description: |
            Тема оформления UI: light, dark или system (как на устройстве / в ОС).
            По умолчанию у новых пользователей — system.
...
    AdminSettings:
      type: object
      required: [external_url, secret_key_set]
      properties:
        external_url:
          type: string
          nullable: true
        secret_key_set:
          type: boolean
          description: true, если `notification_secret_key` задан в БД
    AdminSettingsUpdate:
      type: object
      required: [external_url]
      properties:
        external_url:
          type: string
    FeatureFlagsSnapshot:
      type: object
      description: |
        Снимок инстанс-флагов модулей (`key` → enabled).
        Ключи: registration, debts, credits, budget, subscriptions, recurring,
        balance_maintenance, import_export, stats, notifications, merchants_tags,
        transaction_templates.
      additionalProperties:
        type: boolean
      example:
        budget: true
        registration: false
    AdminFeatureFlag:
      type: object
      required: [key, enabled, title_key, description_key, default_enabled]
      properties:
        key:
          type: string
        enabled:
          type: boolean
        title_key:
          type: string
          description: Ключ i18n для названия модуля
        description_key:
          type: string
          description: Ключ i18n для описания
        default_enabled:
          type: boolean
    AdminFeatureFlagsUpdate:
      type: object
      description: Частичный или полный набор overrides (`key` → enabled). Неизвестный ключ — 400.
      additionalProperties:
        type: boolean
      example:
        budget: false
        registration: true
...
    NameCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
    TransactionTemplate:
      type: object
      required: [id, name, type, sort_order, tags, created_at, updated_at]
      description: Шаблон операции (пресет полей для формы создания)
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        account_id:
          type: string
          nullable: true
        to_account_id:
          type: string
          nullable: true
          description: Только для transfer
        category_id:
          type: string
          nullable: true
        subcategory_id:
          type: string
          nullable: true
        amount:
          type: integer
          nullable: true
          description: Копейки; null — ввести при создании операции
        description:
          type: string
          nullable: true
        merchant_id:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        sort_order:
          type: integer
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagRef'
        created_at:
          type: string
        updated_at:
          type: string
    TransactionTemplateUpsert:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          maxLength: 80
        type:
          type: string
          enum: [income, expense, transfer]
        account_id:
          type: string
          nullable: true
        to_account_id:
          type: string
          nullable: true
        category_id:
          type: string
          nullable: true
        subcategory_id:
          type: string
          nullable: true
        amount:
          type: integer
          nullable: true
          description: Копейки; не передавать или null — сумма при создании
        description:
          type: string
          nullable: true
        merchant_id:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        tag_ids:
          type: array
          items:
            type: string
    Merchant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
          description: ID иконки из каталога категорий (`data/category_icons.json`)
        created_at:
          type: string
    MerchantUpsert:
      type: object
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: string
          description: ID иконки; пусто при создании → `default`, при обновлении — не менять
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
    TagRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    ImportReport:
      type: object
      properties:
        total_rows:
          type: integer
        processed_rows:
          type: integer
          description: Сколько строк уже обработано (для async progress)
        valid_rows:
          type: integer
        skipped_duplicates:
          type: integer
        created_transactions:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              message:
                type: string
        logs:
          type: array
          description: Построчный лог импорта (используется в UI-спойлере полного лога)
          items:
            type: string
        preview:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              action:
                type: string
              account:
                type: string
              amount:
                type: integer
              date:
                type: string
        accounts_to_create:
          type: array
          items:
            type: string
        account_mappings:
          type: array
          items:
            type: object
            properties:
              file_name:
                type: string
              mode:
                type: string
                enum: [create, existing]
              account_id:
                type: string
              account_name:
                type: string
              account_type:
                type: string
                enum: [cash, bank, credit_card]
                description: Предлагаемый тип при mode=create
              bank_id:
                type: string
                description: ID банка при account_type=bank или credit_card
              credit_limit:
                type: string
                description: Лимит карты, рубли; обязателен при account_type=credit_card и mode=create
        categories_to_create:
          type: array
          items:
            type: string
        category_mappings:
          type: array
          items:
            type: object
            properties:
              file_name:
                type: string
              type:
                type: string
                enum: [expense, income]
              mode:
                type: string
                enum: [create, existing]
              category_id:
                type: string
              category_name:
                type: string
        subcategory_mappings:
          type: array
          items:
            type: object
            properties:
              file_category:
                type: string
              file_subcategory:
                type: string
              type:
                type: string
                enum: [expense, income]
              mode:
                type: string
                enum: [create, existing]
              subcategory_id:
                type: string
              subcategory_name:
                type: string
    ImportJob:
      type: object
      required: [id, filename, status, created_at]
      properties:
        id:
          type: string
        filename:
          type: string
        status:
          type: string
          enum: [queued, running, done, failed]
        error_message:
          type: string
          nullable: true
        report:
          $ref: '#/components/schemas/ImportReport'
        created_at:
          type: string
        started_at:
          type: string
          nullable: true
        finished_at:
          type: string
          nullable: true
...